+++ lib/features/settings/presentation/widgets/change_passcode_flow.dart
@@ -1,5 +1,5 @@
 import 'package:coffeecard/base/strings.dart';
-import 'package:coffeecard/widgets/components/forms/settings/change_passcode_form.dart';
+import 'package:coffeecard/features/settings/presentation/widgets/forms/change_passcode_form.dart';
 import 'package:coffeecard/widgets/components/scaffold.dart';
 import 'package:coffeecard/widgets/routers/app_flow.dart';
 import 'package:flutter/material.dart';
